entrega de alguns desafios e endpoints mínimos

+++ semana18/cookenu/src/endpoints/deleteRecipe.ts
@@ -11,31 +11,33 @@ export default async function deleteRecipe(
 
     const token = req.headers.authorization as string;
     const authenticationData  = getTokenData(token);
-    const {id} = req.query;
+    const {id} = req.params;
+
+    if(!id){
+       throw new Error("Id não encontrado!");
+    };
 
     if(authenticationData.role === USER_ROLES.ADMIN) {
-        const userAdmin = await connection("cookenu_recipes")
+        await connection("cookenu_recipes")
         .delete()
         .where({id});
         res.statusCode = 200;
         res.statusMessage = "Receita(s) deletada(s) com sucesso";
     };
 
-    // if(authenticationData.role !== USER_ROLES.ADMIN) {
-    //     res.statusCode = 403;
-    //     res.statusMessage = "Somente usuários com role ADMIN podem deletar receitas sem sua autoria"
-    //     throw new Error()
-    //  };
-
     if(!authenticationData) {
         throw new Error("Esta receita não foi de sua autoria. Favor, editar receitas criadas por você")
     };
 
-       const userNormal = await connection("cookenu_recipes")
-        .delete()
-        .where({user_id: authenticationData.id});
+    if(authenticationData.role === USER_ROLES.NORMAL){
+      await connection("cookenu_recipes")
+      .delete()
+      .where({user_id: authenticationData.id});
+      res.statusCode = 200;
+      res.statusMessage = "Sua receita foi deletada com sucesso!" 
+    };
 
-      res.status(202).send("Sua receita foi deletada com sucesso!");
+      res.end();
 
    } catch (error) {
       res.status(400).send({ message: error.message })
